complete the functions(to be tested)

+++ app/src/main/java/xsf_cym/culife/Bus.java
@@ -181,10 +181,10 @@ public class Bus implements Serializable {
         
     }
 
-    public int estimateTime(String start, String end){
+    public double estimateTime(String start, String end){
         int startIndex = passStops.indexOf(start);
         int endIndex = passStops.indexOf(end);
-        int sum = 0;
+        double sum = 0;
         for(int i = startIndex; i < endIndex; i++){
             sum += interval.get(i);
         }
